+++ packages/core/echo/echo-schema/src/query.test.ts
@@ -5,13 +5,13 @@
 import { expect } from 'chai';
 
 import { sleep } from '@dxos/async';
-import { QUERY_ALL_MODELS, ShowDeletedOption } from '@dxos/echo-db';
 import { beforeAll, beforeEach, describe, test } from '@dxos/test';
 
 import { type EchoDatabase } from './database';
 import { TestBuilder, createDatabase } from './testing';
 import { Text } from './text-object';
 import { Expando, TypedObject } from './typed-object';
+import { QUERY_ALL_MODELS, ShowDeletedOption } from './filter';
 
 describe('Queries', () => {
   let db: EchoDatabase;
